Implement a button to create an RMap DiSCO for the current OSF entity.
This submission adds a "Create RMap DiSCO" button to the title bar of
the RMap widget
that was introduced to the OSF web site in the prevous submission.
Clicking the button
raises a modal dialog warning the users of the consequences of this
action and asking
them to confirm the decision.  Confirmation causes a server request to
be made that
will create the DiSCO and displays the resulting graph in the widget.

The profile.mako and project.mako files now pass arguments to the widget
template
that indicate whether the current node is for a user, a registration, or
a private
project/component.  The node's unique ID is also determined during the
Mako
pre-processing and is passed to the widget.

When a DiSCO is to be created, the JavaScript code calls a new local
server, run on
port 7000.  This server provides the necessary credentials when calling
the RMap
server to request the DiSCO creation.  That call returns the new DiSCO's
ID, which in
turn is returned to JavaScript for use when displaying the graph.  The
new server
must be started along with all the other servers and processes required
by the OSF
site.  On Linux, it is started simply with:

cd osf.io/rmap
./rmap_cgi_server.py &

In order for the RMap server to call back to the OSF API server, it was
changed to
listen to requests from all IP addresses.  The OSF web server was
changed to listen
to the local IP address 127.0.0.1.  It is likely that these values could
be modified
to provide a more secure site.

The modal confirmation dialog was derived from similar Bootstrap code
found elsewhere
in the OSF site.  To overcome a zooming problem in the widget when it
was collapsed
and expanded again, event handlers were added to hide the widget during
the animated
collapse/expand operations.