Roll back some changes based on pull request review.
Most of the changes to __init__.py that were included in the last pull
request were rolled back because they are not needed when viewing the
OSF site in a browser on the same machine as the server, which is how it
is currently being used.  The one change that remains is to allow the
RMap server to call the OSF server to get data it needs to create a
DiSCO.

Also adds a couple of comment lines that somehow escaped the last
submission.